backend/requirements.txt
# Core dependencies
Django==5.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg[binary]==3.2.9
dj-database-url==2.1.0

# Authentication & Security
django-cors-headers==4.3.1
python-jose==3.3.0
python-decouple==3.8

# File handling
Pillow==10.4.0
django-cleanup==8.0.0

# API Documentation
drf-yasg==1.21.7

# Development
django-debug-toolbar==4.3.0

# Environment variables
python-dotenv==1.0.0
django-environ==0.11.2

# Payment processing
stripe==7.11.0

# Data validation and filtering
django-filter==23.5

# Email
sendgrid==6.11.0
